Перенёс метод вычисления длинны названия категории в отдельный вспомогательный класс
