+++ TrackerVC/BaseModel/BaseModelViewController.swift
@@ -2,38 +2,6 @@ import UIKit
 
 class BaseModelViewController: UIViewController {
     
-    enum placeholderText {
-        case tracker
-        case category
-        
-        var text: String {
-            return switch self {
-            case .tracker: "Введите название трекера"
-            case .category: "Введите название категории"
-            }
-        }
-    }
-    
-    enum buttonName {
-        case habit
-        case irregular
-        case create
-        case ready
-        case cancel
-        case add
-        
-        var text: String {
-            return switch self {
-            case .habit: "Привычка"
-            case .irregular: "Нерегулярное событие"
-            case .create: "Создать"
-            case .ready: "Готово"
-            case .cancel: "Отменить"
-            case .add: "Добавить категорию"
-            }
-        }
-    }
-    
     func madeTextField(placeholder: placeholderText) -> UITextField {
         let textField = UITextField()
         textField.placeholder = placeholder.text
@@ -123,3 +91,4 @@ class BaseModelViewController: UIViewController {
         }
     }
 }
+
